Lucky Number
This is a Program where a random number is picked every time the program runs and two users try to find this number. Then, depending on the try the user finds the lucky number (if he finds it), the Program finds a winner. The Program accepts five attempts for each user and then checks who find it faster.