Citizen Registry
This Program implements the idea of having a "registry machine". Simply the program asks for and receives data for a citizen and the user has to input the appropriate information on each field, therefore the user can end the process of adding citizens and at the end of the program, the user has the ability to print out on his screen the citizens registered along with their information.